requirements.txt
chromadb~=0.4.13
langchain==0.0.304
numpy~=1.25
pydantic~=2.4
python-dotenv~=1.0
sentence_transformers~=2.2
